# Conflicts:
#	internal/comet/server_ws_dispatch.go

+++ internal/comet/conf/conf.go
@@ -120,9 +120,9 @@ type Protocol struct {
 }
 
 type Bucket struct {
-	Size          int    `yaml:"size"`
-	Channel       int    `yaml:"channel"`
-	Room          int    `yaml:"room"`
-	RoutineAmount uint64 `yaml:"routineAmount"`
-	RoutineSize   int    `yaml:"routineSize"`
+	Size          int `yaml:"size"`
+	Channel       int `yaml:"channel"`
+	Room          int `yaml:"room"`
+	RoutineAmount int `yaml:"routineAmount"`
+	RoutineSize   int `yaml:"routineSize"`
 }
